formatted + some small changes
