+++ src/services/auth/auth-router.ts
@@ -1,41 +1,22 @@
 import { Router } from "express";
-import { Config } from "../../config";
 import { StatusCodes } from "http-status-codes";
 import { Database } from "../../database";
 import RoleChecker from "../../middleware/role-checker";
 import { Role } from "../auth/auth-models";
-import { AuthLoginValidator, AuthRoleChangeRequest, PlatformValidator } from "./auth-schema";
+import {
+    AuthLoginValidator,
+    AuthRoleChangeRequest,
+    PlatformValidator,
+} from "./auth-schema";
 import authSponsorRouter from "./sponsor/sponsor-router";
 import { CorporateDeleteRequest, CorporateValidator } from "./corporate-schema";
-import { OAuth2Client } from "google-auth-library";
-import { generateJWT, updateDatabaseWithAuthPayload } from "./auth-utils";
-
-const createOAuthClient = (clientId: string, clientSecret?: string) => {
-    return new OAuth2Client({
-        clientId,
-        clientSecret,
-    });
-};
-
-const oauthClients = {
-    web: createOAuthClient(Config.CLIENT_ID, Config.CLIENT_SECRET),
-    ios: createOAuthClient(Config.IOS_CLIENT_ID),
-    // android: createOAuthClient(Config.ANDROID_CLIENT_ID),
-};
-
-
-const getOAuthClient = (platform: string) => {
-    const client = oauthClients[platform as keyof typeof oauthClients];
-    if (!client) {
-        console.warn(`Unknown platform: ${platform}, falling back to web client`);
-        return oauthClients.web;
-    }
-
-    return client;
-};
+import {
+    generateJWT,
+    getOAuthClient,
+    updateDatabaseWithAuthPayload,
+} from "./auth-utils";
 
 const authRouter = Router();
-
 
 authRouter.use("/sponsor", authSponsorRouter);
 
@@ -74,9 +55,9 @@ authRouter.put("/", RoleChecker([Role.Enum.ADMIN]), async (req, res) => {
 });
 
 const getAuthPayloadFromCode = async (
-    code: string, 
-    redirect_uri: string, 
-    platform: string = 'web',
+    code: string,
+    redirect_uri: string,
+    platform: string = "web",
     code_verifier?: string
 ) => {
     try {
@@ -106,45 +87,53 @@ const getAuthPayloadFromCode = async (
 
 authRouter.post("/login/:platform", async (req, res) => {
     try {
-        const { code, redirectUri, codeVerifier } = AuthLoginValidator.parse(req.body);
+        const { code, redirectUri, codeVerifier } = AuthLoginValidator.parse(
+            req.body
+        );
         const platform = PlatformValidator.parse(req.params.platform);
-        
-        const authPayload = await getAuthPayloadFromCode(code, redirectUri, platform, codeVerifier);
 
-    if (!authPayload) {
-        return res
-            .status(StatusCodes.BAD_REQUEST)
-            .send({ error: "InvalidToken" });
-    }
+        const authPayload = await getAuthPayloadFromCode(
+            code,
+            redirectUri,
+            platform,
+            codeVerifier
+        );
 
-    const properScopes =
-        "email" in authPayload && "sub" in authPayload && "name" in authPayload;
-    if (!properScopes) {
-        return res
-            .status(StatusCodes.BAD_REQUEST)
-            .send({ error: "InvalidScopes" });
-    }
+        if (!authPayload) {
+            return res
+                .status(StatusCodes.BAD_REQUEST)
+                .send({ error: "InvalidToken" });
+        }
 
-    // Update database by payload
-    await updateDatabaseWithAuthPayload(authPayload);
+        const properScopes =
+            "email" in authPayload &&
+            "sub" in authPayload &&
+            "name" in authPayload;
+        if (!properScopes) {
+            return res
+                .status(StatusCodes.BAD_REQUEST)
+                .send({ error: "InvalidScopes" });
+        }
 
-    // Generate the JWT
-    const jwtToken = await generateJWT(`user${authPayload.sub}`);
+        // Update database by payload
+        await updateDatabaseWithAuthPayload(authPayload);
 
-    return res.status(StatusCodes.OK).send({ token: jwtToken });
+        // Generate the JWT
+        const jwtToken = await generateJWT(`user${authPayload.sub}`);
+
+        return res.status(StatusCodes.OK).send({ token: jwtToken });
     } catch (error) {
         console.error("Error in platform login:", error);
-        return res.status(StatusCodes.BAD_REQUEST).send({ 
+        return res.status(StatusCodes.BAD_REQUEST).send({
             error: "InvalidRequest",
-            details: error instanceof Error ? error.message : "Unknown error"
+            details: error instanceof Error ? error.message : "Unknown error",
         });
     }
 });
 
-
 authRouter.post("/login/", async (req, res) => {
     const { code, redirectUri } = AuthLoginValidator.parse(req.body);
-    const authPayload = await getAuthPayloadFromCode(code, redirectUri, 'web');
+    const authPayload = await getAuthPayloadFromCode(code, redirectUri, "web");
 
     if (!authPayload) {
         return res
