+++ src/services/auth/auth-utils.ts
@@ -1,9 +1,33 @@
 // Create a function to generate GoogleStrategy instances
-import { TokenPayload } from "google-auth-library";
+import { OAuth2Client, TokenPayload } from "google-auth-library";
 import { Config } from "../../config";
 import { Database } from "../../database";
 import { JwtPayloadType, Role } from "./auth-models";
 import jsonwebtoken from "jsonwebtoken";
+
+const createOAuthClient = (clientId: string, clientSecret?: string) => {
+    return new OAuth2Client({
+        clientId,
+        clientSecret,
+    });
+};
+
+const oauthClients = {
+    web: createOAuthClient(Config.CLIENT_ID, Config.CLIENT_SECRET),
+    ios: createOAuthClient(Config.IOS_CLIENT_ID),
+};
+
+export const getOAuthClient = (platform: string) => {
+    const client = oauthClients[platform as keyof typeof oauthClients];
+    if (!client) {
+        console.warn(
+            `Unknown platform: ${platform}, falling back to web client`
+        );
+        return oauthClients.web;
+    }
+
+    return client;
+};
 
 export async function updateDatabaseWithAuthPayload(payload: TokenPayload) {
     const userId = `user${payload.sub}`;
